improve code modularity
