+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.2] - 2024-10-16
+### Changed
+- Split up code in Neurosynchrony into several modules to improve separation of concerns and code maintenance.
+
+
 ## [2.5.1] - 2024-10-15
 ### Fixed
 - Protocol update logic, protocol can only be updated between 8-13 Hz.
